+++ code/待整理/c13.02/source/comm/cpu_instr.h
@@ -48,7 +48,7 @@ static inline void lgdt(uint32_t start, uint32_t size) {
 	__asm__ __volatile__("lgdt %0"::"m"(gdt));
 }
 
-static inline uint16_t read_cr0() {
+static inline uint32_t read_cr0() {
 	uint32_t cr0;
 	__asm__ __volatile__("mov %%cr0, %0":"=r"(cr0));
 	return cr0;
